% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_station_names.R
\name{clean_station_names}
\alias{clean_station_names}
\title{Clean and Standardize Station Names}
\usage{
clean_station_names(name)
}
\arguments{
\item{name}{A character vector of station names to be cleaned.}
}
\value{
A character vector of cleaned station names with standardized formatting.
}
\description{
This function cleans and standardizes station names by performing several transformations:
}
\examples{
stations <- c("Mississippi Sound near Grand Pass", "Rigolets at Hwy 90 near
Slidell, LA", "EAST PEARL RIVER AT CSX RAILROAD NR CLAIBORNE, MS",
"MISSISSIPPI SOUND AT USGS MERRILL SHELL BANK LIGHT", "MISSISSIPPI SOUND AT
USGS ROUND ISLAND LIGHT, MS", "BILOXI BAY AT POINT CADET HARBOR AT BILOXI, MS")
clean_station_names(stations)

}
