% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDEQ_beach_stations.R
\docType{data}
\name{MDEQ_beach_stations}
\alias{MDEQ_beach_stations}
\title{MDEQ Beach Stations}
\format{
An \code{sf} object, using CRS NAD83 (EPSG:4269) with 21 features (rows) and 3
fields (columns), with point geometry for each.
\describe{
\item{\code{DEQStationID}}{character, station ID matching that on DEQ's beach
website. In WQX, this station is provided as part of the
'MonitoringLocationDescriptionText' field.}
\item{\code{WQXStationID}}{character, station ID matching that in the
'MonitoringLocationIdentifier field from a Water Quality Portal download.}
\item{\code{StationName}}{character, station name as converted to title case from
all caps in the 'MonitoringLocationName' field of a Water Quality Portal
download. This corresponds to the station names as given on the DEQ beach
website, with some minor differences that do not affect station identification
(e.g. 'Long Beach' here is 'Long Beach Beach' on the beach website).}
\item{\code{geometry}}{POINT, degrees longitude and latitude, obtained from WQX.}
}
}
\source{
Combination of data download from a \href{https://www.waterqualitydata.us/#organization=21MSWQ_WQX&project=BCH24&mimeType=csv&providers=NWIS&providers=STORET}{query of stations in the BCH24 project}
in the Water Quality Portal and cross-checking with stations on the \href{https://beaches.mdeq.ms.gov/}{MDEQ beach website}.
}
\usage{
MDEQ_beach_stations
}
\description{
Station information for MDEQ's recreational beach sampling program
}
\details{
The Mississippi Department of Environmental Quality (MDEQ) regularly samples
21 stations across the MS coast for Enterococcus, a fecal indicator bacteria,
as well as various nutrients, chlorophyll \emph{a}, and Total Suspended Solids.
Data is available through their \href{https://beaches.mdeq.ms.gov/}{beach website}
and through the \href{https://www.waterqualitydata.us/}{Water Quality Portal},
using Organization ID '21MSWQ_WQX' and Project ID 'BCHYY', where 'YY' is the
last two digits of the year of interest.
}
\keyword{datasets}
