% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msep_boundary.R
\docType{data}
\name{msep_boundary}
\alias{msep_boundary}
\title{Mississippi Sound Watershed Boundary}
\format{
## \code{msep_boundary}
An \code{sf} object, using CRS NAD83 (EPSG:4269), with a single polygon representing
the MSEP outline.
}
\usage{
msep_boundary
}
\description{
Outline of the Mississippi Sound Watershed, which is the combination of
HUC4s 0317 and 0318.
}
\details{
NOT INTENDED FOR ANALYTICAL USE. The boundary has been simplified
to reduce object size. For accurate boundaries, and other possible boundary
configurations (e.g. only the portion of the watershed that is inside the
state of MS), see the \code{msepBoundaries} package: \url{https://github.com/CMEP-MS/msepBoundaries}
}
\keyword{datasets}
