% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_hydro.R
\name{mapview_hydro}
\alias{mapview_hydro}
\title{Interactive mapview plot with USGS Hydrography overlay}
\usage{
mapview_hydro(x, map.types = "default", layer.name = "data", ...)
}
\arguments{
\item{x}{Spatial object (e.g., \code{sf}, \code{SpatVector}).}

\item{map.types}{Character vector of basemaps. Defaults to \code{c("Esri.OceanBasemap", "Esri.WorldGrayCanvas", "Esri.WorldTopoMap", "OpenStreetMap")} via an internal function.}

\item{layer.name}{Character name for data layer. Defaults to "data".}

\item{...}{Additional arguments passed to \code{\link[mapview:mapView]{mapview::mapview()}}.}
}
\value{
A \code{mapview} object containing the modified leaflet map.
}
\description{
Basemap options within common R spatial packages are not entirely satisfactory when
mapping along estuarine coastlines. USGS provides a Hydrography layer that is
very useful, but is not built into said packages. This function adds the USGS
Hydrography layer to a mapview object. This function can be used just like
\code{mapview::mapview()}.
}
\examples{
\dontrun{
library(mapview)
# make a map as you would otherwise with mapview
# styling is adjusted automatically
mapview_hydro(MDEQ_beach_stations,
              color = "purple",
              lwd = 2,
              col.regions = "orange",
              legend = NULL)
# give it a layer name
mapview_hydro(MDEQ_beach_stations,
              layer.name = "MDEQ Beach Stations",
              color = "purple",
              lwd = 2,
              col.regions = "orange")
# provide different basemaps
mapview_hydro(MDEQ_beach_stations,
              color = "purple",
              col.regions = "orange",
              legend = NULL,
              map.types = c("Esri.WorldTopoMap",
                            "OpenStreetMap"))
}

}
