% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msep_boundary.R
\docType{data}
\name{msep_boundary}
\alias{msep_boundary}
\title{Mississippi Sound Estuary Program Boundary}
\format{
## `msep_boundary`
An `sf` object, using CRS NAD83 (EPSG:4269), with a single polygon representing
the MSEP outline.
}
\usage{
msep_boundary
}
\description{
Outline of the MSEP boundary in the state of MS. Originally generated using
files from Mississippi Department of Environmental Quality, cropped with
the state boundary from the `\{rnaturalearth\}` package. Provided for
assistance with visualization, not analytical use.
}
\keyword{datasets}
