% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_hydro.R
\name{tmap_hydro}
\alias{tmap_hydro}
\title{Add USGS Hydrography overlay to a tmap object}
\usage{
tmap_hydro(tm_obj, base_groups = hydro_basemaps("tmap"))
}
\arguments{
\item{tm_obj}{A \code{tmap} object. \strong{Note}, do not add a \code{tm_basemap} layer if using plot mode - it will error.}

\item{base_groups}{Character vector of basemaps. Defaults to \verb{\\r hydro_basemaps("tmap")}.}
}
\value{
A map object: in plot mode, a regular \code{tmap} object; in view mode,
a \code{leaflet} map object.
}
\description{
Basemap options within common R spatial packages are not entirely satisfactory when
mapping along estuarine coastlines. USGS provides a Hydrography layer that is
very useful, but is not built into said packages. This function adds the USGS
Hydrography layer to a tmap object that has already been created. In view mode,
the output is turned into an interactive leaflet object. In plot mode, it
remains a tmap object. This function works with vector objects but is currently untested on rasters.
}
\examples{
\dontrun{
library(tmap)
# make a map as desired with tmap
tm <- tm_shape(MDEQ_beach_stations) +
    tm_symbols(fill = "blue")

# add the tmap_hydro styling to the tmap object
# whatever mode tmap happens to be in
tmap_hydro(tm)

# plot mode specifically
tmap_mode("plot")
tmap_hydro(tm)

# view mode specifically
# this provides multiple base layers that can be chosen
# and you can turn off the USGS Hydro layer if you want to
tmap_mode("view")
tmap_hydro(tm)
}

}
